chessapp/requirements.txt
django
django-heroku
python-chess